Check if bookmarked before updating bookmarks
