+++ src/lib/bookmarks.ts
@@ -59,3 +59,18 @@ export function getButtonInfo(bookmark: Bookmark) {
         buttonColor,
     };
 }
+
+export async function checkIfBookmarked(mangaId: string) {
+    const user_data = localStorage.getItem("accountInfo");
+
+    if (!user_data) {
+        console.error("User data not found");
+        return false;
+    }
+
+    const response = await fetch(
+        `/api/bookmarks/${mangaId}?user_data=${encodeURIComponent(user_data)}`,
+    );
+    const data = await response.json();
+    return data.isBookmarked;
+}
